Move KV cache into Week 2
Teach and test dense incremental decoding before kernel optimization, then start Week 3 with continuous batching and paged-cache serving work.
