+++ book/src/SUMMARY.md
@@ -15,17 +15,17 @@
     - [Sampling and Preparing for Week 2](./week1-07-sampling-prepare.md)
 - [Week 2: A Step Closer to vLLM](./week2-overview.md)
     - [🚧 Benchmark MLX](./week2-01-benchmarking.md)
-    - [Quantized Matmul](./week2-02-quantized-matmul.md)
-    - [🚧 Fast Operators](./week2-03-fast-operators.md)
-    - [🚧 Decode Attention](./week2-04-decode-attention.md)
-    - [🚧 End-to-End Optimization](./week2-05-end-to-end.md)
+    - [Key-Value Cache](./week2-02-kv-cache.md)
+    - [Quantized Matmul](./week2-03-quantized-matmul.md)
+    - [🚧 Fast Operators](./week2-04-fast-operators.md)
+    - [🚧 Decode Attention](./week2-05-decode-attention.md)
+    - [🚧 End-to-End Optimization](./week2-06-end-to-end.md)
 - [Week 3: Build a Mini vLLM](./week3-overview.md)
-    - [Key-Value Cache](./week3-01-kv-cache.md)
-    - [Continuous Batching](./week3-02-continuous-batching.md)
-    - [FlashAttention for Prefill](./week3-03-flash-attention.md)
-    - [Chunked Prefill](./week3-04-chunked-prefill.md)
-    - [Paged Attention, Part 1](./week3-05-paged-attention-part1.md)
-    - [Paged Attention, Part 2](./week3-06-paged-attention-part2.md)
+    - [Continuous Batching](./week3-01-continuous-batching.md)
+    - [FlashAttention for Prefill](./week3-02-flash-attention.md)
+    - [Chunked Prefill](./week3-03-chunked-prefill.md)
+    - [Paged Attention, Part 1](./week3-04-paged-attention-part1.md)
+    - [Paged Attention, Part 2](./week3-05-paged-attention-part2.md)
     - [Optional: Mixture of Experts](./week3-optional-moe.md)
     - [🚧 Optional: Speculative Decoding](./week3-optional-speculative-decoding.md)
 - [🚧 Week 4: Build a Coding Agent](./week4-overview.md)
