+++ book/src/glossary.md
@@ -14,11 +14,11 @@
 - [Qwen3 Transformer Block](./week1-05-qwen3-model.md)
 - [Week 1 Qwen3 Model](./week1-05-qwen3-model.md)
 - [dequantize_linear](./week1-05-qwen3-model.md)
-- [KV Cache](./week3-01-kv-cache.md)
-- [Quantized Matmul](./week2-02-quantized-matmul.md)
-- [Fast RMSNorm, RoPE, and SwiGLU](./week2-03-fast-operators.md)
-- [Decode Attention](./week2-04-decode-attention.md)
-- [Flash Attention](./week3-03-flash-attention.md)
-- [Paged Attention](./week3-06-paged-attention-part2.md)
+- [KV Cache](./week2-02-kv-cache.md)
+- [Quantized Matmul](./week2-03-quantized-matmul.md)
+- [Fast RMSNorm, RoPE, and SwiGLU](./week2-04-fast-operators.md)
+- [Decode Attention](./week2-05-decode-attention.md)
+- [Flash Attention](./week3-02-flash-attention.md)
+- [Paged Attention](./week3-05-paged-attention-part2.md)
 
 {{#include copyright.md}}
