+++ book/src/week2-01-benchmarking.md
@@ -20,6 +20,12 @@ phase while hurting the other, so `bench.py` reports both:
 The first generated token belongs to prefill. Excluding it from decode prevents
 prompt length from distorting the decode number.
 
+Both sides of the Week 2 comparison must use a KV cache: prefill the prompt
+once, then pass only the newly generated token on each decode step. Comparing a
+cached MLX baseline with a course model that recomputes the full prefix would
+measure two different algorithms and make the kernel target meaningless. Day 2
+implements the dense cache used by every later Week 2 benchmark.
+
 ## Synchronize Lazy Work
 
 MLX builds lazy computation graphs. Timing only the Python call measures graph
