+++ book/src/week2-06-end-to-end.md
@@ -1,4 +1,4 @@
-# 🚧 Week 2 Days 6-7: End-to-End Optimization
+# 🚧 Week 2 Day 7: End-to-End Optimization
 
 > 🚧 This newly introduced chapter is a work in progress.
 
@@ -24,8 +24,8 @@ positions, including correctness tests and prompt scoring, pass `None`.
 
 - `qwen3_week1.py` keeps readable Python operators and dequantized weights.
 - `qwen3_week2.py` uses quantized weights and `week2_kernels.py`.
-- `qwen3_week3.py` imports Week 2 model components, then adds serving caches and
-  paged attention.
+- `qwen3_week3.py` imports Week 2 model components, reuses the cache interface,
+  and adds paged cache storage and serving mechanisms.
 
 Later weeks must not overwrite earlier files. The starter and reference trees
 expose the same module and function names so each checkpoint remains runnable.
@@ -33,7 +33,7 @@ expose the same module and function names so each checkpoint remains runnable.
 ## Validate Correctness
 
 ```bash
-pdm run test --week 2 --day 6
+pdm run test --week 2 --day 7
 ```
 
 Check full-sequence logits with `logits_to_keep=None`, last-position logits with
