+++ book/src/week3-03-chunked-prefill.md
@@ -1,4 +1,4 @@
-# Week 3 Day 4: Chunked Prefill
+# Week 3 Day 3: Chunked Prefill
 
 A long prompt can monopolize the device while active decode requests wait for
 their next token. Chunked prefill limits the number of prompt tokens admitted in
