+++ tests_refsol/test_week_2_day_3.py
@@ -1,0 +1,147 @@
+"""Week 2 Days 3-4 quantized-operator tests."""
+
+import inspect
+
+import mlx.core as mx
+
+import tiny_llm_ref.embedding as embedding_module
+import tiny_llm_ref.quantize as quantize_module
+from tiny_llm_ref.embedding import QuantizedEmbedding
+from tiny_llm_ref.quantize import QuantizedWeights
+
+from .tiny_llm_base import quantized_matmul, quantized_matvec_custom
+from .utils import assert_allclose
+
+
+def test_task_1_quantized_embedding_dequantizes_selected_rows():
+    weight = mx.random.normal((7, 256)).astype(mx.bfloat16)
+    packed, scales, biases = mx.quantize(weight, group_size=128, bits=4)
+    quantized = QuantizedWeights(scales, biases, 128, 4, packed)
+    embedding = QuantizedEmbedding(7, 256, quantized)
+    indices = mx.array([[1, 4]])
+
+    result = embedding(indices)
+    expected = mx.dequantize(
+        packed[indices], scales[indices], biases[indices], group_size=128, bits=4
+    )
+    assert_allclose(result, expected, mx.bfloat16, atol=2e-2, rtol=2e-2)
+
+
+def test_week2_quantization_path_uses_course_owned_operators():
+    source = (
+        inspect.getsource(quantize_module.quantized_matmul)
+        + inspect.getsource(quantize_module.dequantize_weights)
+        + inspect.getsource(embedding_module.QuantizedEmbedding.__call__)
+    )
+    assert "mx.quantized_matmul" not in source
+    assert "mx.dequantize" not in source
+
+
+def quantized_matmul_helper(
+    stream: mx.Stream,
+    precision: mx.Dtype,
+    identity_matrix: bool,
+):
+    with mx.stream(stream):
+        group_size = 128
+        if identity_matrix:
+            input = mx.eye(group_size, dtype=precision)
+        else:
+            input = mx.random.normal(shape=(3, group_size), dtype=precision)
+        weight = mx.random.normal(shape=(5, group_size), dtype=precision)
+        w_q, scales, biases = mx.quantize(weight, group_size=group_size, bits=4)
+        user_out = quantized_matmul(
+            scales=scales,
+            biases=biases,
+            group_size=group_size,
+            bits=4,
+            a=input,
+            b=w_q,
+            transpose_b=True,
+        )
+        ref_out = mx.quantized_matmul(
+            input,
+            w_q,
+            scales,
+            biases,
+            group_size=group_size,
+            bits=4,
+            transpose=True,
+        )
+        if identity_matrix:
+            assert_allclose(user_out, ref_out, precision)
+        else:
+            assert_allclose(
+                user_out,
+                ref_out,
+                precision,
+                atol=5.0e-1,
+                message=f"quantized matmul {precision} comparison",
+            )
+
+
+def test_task_2_quantized_matmul_simple_bf16_cpu():
+    quantized_matmul_helper(mx.cpu, mx.bfloat16, True)
+
+
+def test_task_2_quantized_matmul_complex_bf16_cpu():
+    quantized_matmul_helper(mx.cpu, mx.bfloat16, False)
+
+
+def test_task_2_quantized_matmul_simple_f16_cpu():
+    quantized_matmul_helper(mx.cpu, mx.float16, True)
+
+
+def test_task_2_quantized_matmul_complex_f16_cpu():
+    quantized_matmul_helper(mx.cpu, mx.float16, False)
+
+
+def test_task_3_quantized_matmul_simple_bf16_gpu():
+    quantized_matmul_helper(mx.gpu, mx.bfloat16, True)
+
+
+def test_task_3_quantized_matmul_complex_bf16_gpu():
+    quantized_matmul_helper(mx.gpu, mx.bfloat16, False)
+
+
+def test_task_3_quantized_matmul_simple_f16_gpu():
+    quantized_matmul_helper(mx.gpu, mx.float16, True)
+
+
+def test_task_3_quantized_matmul_complex_f16_gpu():
+    quantized_matmul_helper(mx.gpu, mx.float16, False)
+
+
+def quantized_matvec_custom_helper(num_rows: int):
+    with mx.stream(mx.gpu):
+        group_size = 128
+        input = mx.random.normal(shape=(num_rows, group_size), dtype=mx.bfloat16)
+        weight = mx.random.normal(shape=(64, group_size), dtype=mx.bfloat16)
+        w_q, scales, biases = mx.quantize(weight, group_size=group_size, bits=4)
+        user_out = quantized_matvec_custom(
+            scales=scales,
+            biases=biases,
+            group_size=group_size,
+            bits=4,
+            a=input,
+            b=w_q,
+            transpose_b=True,
+        )
+        ref_out = mx.quantized_matmul(
+            input,
+            w_q,
+            scales,
+            biases,
+            group_size=group_size,
+            bits=4,
+            transpose=True,
+        )
+        assert_allclose(user_out, ref_out, mx.bfloat16, atol=5.0e-1)
+
+
+def test_task_4_quantized_matvec_custom_m1_gpu():
+    quantized_matvec_custom_helper(1)
+
+
+def test_task_4_quantized_matvec_custom_m8_gpu():
+    quantized_matvec_custom_helper(8)
