+++ tests_refsol/test_week_2_day_6.py
@@ -1,137 +1,36 @@
-"""Week 2 end-to-end optimized model tests."""
+"""Week 2 Day 6 decode-attention tests."""
 
 import mlx.core as mx
-import pytest
-from .utils import *
-from .tiny_llm_base import (
-    Qwen3ModelWeek2,
-    Qwen3ModelWeek3,
-    TinyKvFullCache,
+
+from tiny_llm_ref.attention import scaled_dot_product_attention_grouped
+from tiny_llm_ref.week2_kernels import (
+    decode_attention_custom,
+    scaled_dot_product_attention,
 )
-from mlx_lm import load
 
-# TODO: task 1 tests
-
-
-@pytest.mark.skipif(
-    not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
-)
-def test_utils_qwen3_0_6b():
-    pass
+from .utils import assert_allclose
 
 
-@pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
-def test_utils_qwen3_4b():
-    pass
-
-
-@pytest.mark.skipif(
-    not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
-)
-def test_utils_qwen3_1_7b():
-    pass
-
-
-def helper_test_task_3(model_name: str, iters: int = 10):
-    mlx_model, tokenizer = load(model_name)
-    model = Qwen3ModelWeek2(mlx_model)
-    for _ in range(iters):
-        cache = [TinyKvFullCache() for _ in range(model.num_hidden_layers)]
-        input = mx.random.randint(low=0, high=tokenizer.vocab_size, shape=(1, 10))
-        user_output = model(input, 0, cache)
-        ref_output = mlx_model(input)
-        user_output = user_output - mx.logsumexp(user_output, keepdims=True)
-        ref_output = ref_output - mx.logsumexp(ref_output, keepdims=True)
-        assert_allclose(
-            user_output, ref_output, precision=mx.bfloat16, rtol=0.1, atol=1.0
-        )
-
-
-@pytest.mark.skipif(
-    not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
-)
-def test_task_3_qwen3_0_6b():
-    helper_test_task_3("Qwen/Qwen3-0.6B-MLX-4bit", 5)
-
-
-@pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
-def test_task_3_qwen3_4b():
-    helper_test_task_3("Qwen/Qwen3-4B-MLX-4bit", 1)
-
-
-@pytest.mark.skipif(
-    not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
-)
-def test_task_3_qwen3_1_7b():
-    helper_test_task_3("Qwen/Qwen3-1.7B-MLX-4bit", 3)
-
-
-def helper_test_task_4(
-    model_name: str,
-    seq_len: int,
-    iters: int = 1,
-):
-    mlx_model, tokenizer = load(model_name)
-    model = Qwen3ModelWeek2(mlx_model)
-    for _ in range(iters):
-        inputs = mx.random.randint(0, tokenizer.vocab_size, (1, seq_len))
-        ref_outputs = mlx_model(inputs)
-        decode_cache = [TinyKvFullCache() for _ in range(model.num_hidden_layers)]
-        for offset in range(seq_len):
-            user_out = model(
-                inputs=inputs[:, offset : offset + 1],
-                offset=offset,
-                cache=decode_cache,
-            )
-            ref_out = ref_outputs[:, offset : offset + 1, :]
-            user_out = user_out - mx.logsumexp(user_out, keepdims=True)
-            ref_out = ref_out - mx.logsumexp(ref_out, keepdims=True)
-            assert_allclose(
-                user_out, ref_out, precision=mx.bfloat16, rtol=0.1, atol=1.0
-            )
-
-
-@pytest.mark.skipif(
-    not qwen3_0_6b_model_exists(), reason="Qwen3-0.6B-4bit model not found"
-)
-def test_task_4_qwen3_0_6b():
-    helper_test_task_4("Qwen/Qwen3-0.6B-MLX-4bit", seq_len=3)
-
-
-@pytest.mark.skipif(not qwen3_4b_model_exists(), reason="Qwen3-4B-4bit model not found")
-def test_task_4_qwen3_4b():
-    helper_test_task_4(
-        "Qwen/Qwen3-4B-MLX-4bit",
-        seq_len=3,
+def test_fast_attention_matches_grouped_attention():
+    query = mx.random.normal((2, 4, 3, 16)).astype(mx.float32)
+    key = mx.random.normal((2, 2, 5, 16)).astype(mx.float32)
+    value = mx.random.normal((2, 2, 5, 16)).astype(mx.float32)
+    mask = mx.broadcast_to(
+        mx.array([0, 0, 0, 0, -mx.inf], dtype=mx.float32), (2, 1, 3, 5)
     )
+    scale = 16**-0.5
+    result = scaled_dot_product_attention(query, key, value, scale, mask)
+    expected = scaled_dot_product_attention_grouped(query, key, value, scale, mask)
+    assert result.shape == query.shape
+    assert_allclose(result, expected, mx.float32, atol=1e-5, rtol=1e-5)
 
 
-@pytest.mark.skipif(
-    not qwen3_1_7b_model_exists(), reason="Qwen3-1.7B-4bit model not found"
-)
-def test_task_4_qwen3_1_7b():
-    helper_test_task_4("Qwen/Qwen3-1.7B-MLX-4bit", seq_len=3)
-
-
-class FakeEmbedding:
-    def __call__(self, inputs):
-        return mx.stack([inputs, inputs + 1], axis=-1).astype(mx.float32)
-
-    def as_linear(self, hidden):
-        return hidden
-
-
-@pytest.mark.parametrize("model_class", [Qwen3ModelWeek2, Qwen3ModelWeek3])
-@pytest.mark.parametrize("logits_to_keep,expected_length", [(1, 1), (None, 4)])
-def test_logits_to_keep_controls_output_length(
-    model_class, logits_to_keep, expected_length
-):
-    model = model_class.__new__(model_class)
-    model.num_hidden_layers = 0
-    model.embedding = FakeEmbedding()
-    model.layers_inner = []
-    model.norm = lambda hidden: hidden
-    model.w_lm_head = None
-    inputs = mx.array([[1, 2, 3, 4]])
-    result = model(inputs, 0, [], logits_to_keep=logits_to_keep)
-    assert result.shape == (1, expected_length, 2)
+def test_custom_metal_attention_matches_grouped_attention():
+    query = mx.random.normal((2, 4, 3, 16)).astype(mx.bfloat16)
+    key = mx.random.normal((2, 2, 5, 16)).astype(mx.bfloat16)
+    value = mx.random.normal((2, 2, 5, 16)).astype(mx.bfloat16)
+    scale = 16**-0.5
+    result = decode_attention_custom(query, key, value, scale, "causal")
+    expected = scaled_dot_product_attention_grouped(query, key, value, scale, "causal")
+    assert result.shape == query.shape
+    assert_allclose(result, expected, mx.bfloat16, atol=2e-2, rtol=2e-2)
