+++ tests_refsol/test_week_3_day_5.py
@@ -1,15 +1,18 @@
-"""Week 3 paged-KV storage tests."""
+"""Week 3 Day 5 paged-attention runtime tests."""
 
 from types import SimpleNamespace
 
 import mlx.core as mx
 
 from .tiny_llm_base import (
+    BatchingKvCache,
     Qwen3ModelWeek2,
     Qwen3ModelWeek3,
-    TinyKvFullCache,
     TinyKvPagedCache,
     TinyKvPagedPool,
+    flash_attention,
+    paged_attention,
+    scaled_dot_product_attention_grouped,
 )
 from .utils import assert_allclose
 
@@ -101,125 +104,93 @@ def _fake_qwen3_mlx_model() -> SimpleNamespace:
     )
 
 
-def test_task_1_paged_cache_matches_full_cache():
+def test_task_1_paged_attention_matches_dense_flash_attention():
     page_size = 4
-    full = TinyKvFullCache()
     pool = TinyKvPagedPool(page_size=page_size)
-    paged = TinyKvPagedCache(pool=pool)
+    cache = TinyKvPagedCache(pool=pool)
+    first_key, first_value = _random_chunk(3)
+    second_key, second_value = _random_chunk(3)
 
-    total_len = 0
-    for length in [3, 2, 5]:
-        key, value = _random_chunk(length)
-        full_key, full_value, full_len, _ = full.update_and_fetch(key, value)
-        paged_key, paged_value, paged_len, _ = paged.update_and_fetch(key, value)
-        total_len += length
-        assert full_len == paged_len == total_len
-        assert paged.num_pages == (total_len + page_size - 1) // page_size
-        physical_page_capacity = [
-            paged.pool.read_page(page_id)[0].shape[2] for page_id in paged.page_ids
-        ]
-        assert physical_page_capacity == [page_size] * paged.num_pages
-        assert sum(paged.page_lens) == total_len
-        assert_allclose(paged_key, full_key, precision=mx.float32)
-        assert_allclose(paged_value, full_value, precision=mx.float32)
+    cache.update_and_fetch(first_key, first_value)
+    metadata = cache.update_and_fetch_paged(second_key, second_value, mask="causal")
+
+    query = mx.random.normal(shape=(1, 4, second_key.shape[2], 4)).astype(mx.float32)
+    dense_key, dense_value = cache.gather_dense()
+    dense_output = flash_attention(
+        query,
+        dense_key,
+        dense_value,
+        mask="causal",
+    )
+    paged_output = paged_attention(
+        query,
+        metadata.key_pages,
+        metadata.value_pages,
+        metadata.block_table,
+        metadata.context_lens,
+        metadata.page_size,
+        mask=metadata.mask,
+    )
+
+    assert metadata.block_table.tolist() == [[0, 1]]
+    assert metadata.context_lens.tolist() == [6]
+    assert metadata.key_pages.shape == (2, 2, page_size, 4)
+    assert metadata.value_pages.shape == (2, 2, page_size, 4)
+    assert_allclose(paged_output, dense_output, precision=mx.float32)
 
 
-def test_task_1_paged_pool_reuses_freed_pages():
-    pool = TinyKvPagedPool(page_size=4)
+def test_task_2_batched_paged_attention_matches_dense_attention():
+    page_size = 4
+    pool = TinyKvPagedPool(page_size=page_size)
     first = TinyKvPagedCache(pool=pool)
     second = TinyKvPagedCache(pool=pool)
+    first.update_and_fetch(*_random_chunk(3))
+    second.update_and_fetch(*_random_chunk(6))
 
-    key, value = _random_chunk(6)
-    first.update_and_fetch(key, value)
-    assert first.page_ids == [0, 1]
-    assert pool.num_pages == 2
-    assert pool.num_free_pages == 0
+    batch = BatchingKvCache(max_active_requests=3, max_seq_len=16)
+    batch.add_request(first, 0)
+    batch.add_request(second, 2)
 
-    first.release()
-    assert first.offset == 0
-    assert pool.num_pages == 2
-    assert pool.num_free_pages == 2
+    keys = mx.zeros((3, 2, 1, 4), dtype=mx.float32)
+    values = mx.zeros((3, 2, 1, 4), dtype=mx.float32)
+    keys[0:1], values[0:1] = _random_chunk(1)
+    keys[2:3], values[2:3] = _random_chunk(1)
 
-    second_key, second_value = _random_chunk(5)
-    gathered_key, gathered_value, seq_len, _ = second.update_and_fetch(
-        second_key, second_value
+    metadata = batch.update_and_fetch_paged(
+        keys,
+        values,
+        mask_length=1,
+        mask="causal",
     )
-    assert seq_len == 5
-    assert pool.num_pages == 2
-    assert pool.num_free_pages == 0
-    assert set(second.page_ids) == {0, 1}
-    assert_allclose(gathered_key, second_key, precision=mx.float32)
-    assert_allclose(gathered_value, second_value, precision=mx.float32)
+    query = mx.random.normal(shape=(3, 4, 1, 4)).astype(mx.float32)
+    paged_output = paged_attention(
+        query,
+        metadata.key_pages,
+        metadata.value_pages,
+        metadata.block_table,
+        metadata.context_lens,
+        metadata.page_size,
+        mask=metadata.mask,
+    )
+
+    first_key, first_value = first.gather_dense()
+    first_output = scaled_dot_product_attention_grouped(
+        query[0:1], first_key, first_value, mask="causal"
+    )
+    second_key, second_value = second.gather_dense()
+    second_output = scaled_dot_product_attention_grouped(
+        query[2:3], second_key, second_value, mask="causal"
+    )
+
+    assert metadata.context_lens.tolist() == [4, 0, 7]
+    assert metadata.block_table.shape == (3, 2)
+    assert metadata.block_table.tolist()[1] == [-1, -1]
+    assert metadata.key_pages.shape == (3, 2, page_size, 4)
+    assert_allclose(paged_output[0:1], first_output, precision=mx.float32)
+    assert_allclose(paged_output[2:3], second_output, precision=mx.float32)
 
 
-def test_task_1_paged_cache_rewind():
-    page_size = 4
-    pool = TinyKvPagedPool(page_size=page_size)
-    paged = TinyKvPagedCache(pool=pool)
-    full = TinyKvFullCache()
-
-    for length in [4, 3, 2]:
-        key, value = _random_chunk(length)
-        paged.update_and_fetch(key, value)
-        full.update_and_fetch(key, value)
-
-    assert paged.page_lens == [4, 4, 1]
-    paged.rewind(3)
-    full.rewind(3)
-
-    paged_key, paged_value = paged.gather_dense()
-    full_key, full_value = full.key_values
-    assert paged.offset == full.offset == 6
-    assert paged.page_lens == [4, 2]
-    assert paged.num_pages == 2
-    assert paged.pool.num_pages == 3
-    assert paged.pool.num_free_pages == 1
-    physical_page_capacity = [
-        paged.pool.read_page(page_id)[0].shape[2] for page_id in paged.page_ids
-    ]
-    assert physical_page_capacity == [page_size] * paged.num_pages
-    assert_allclose(paged_key, full_key, precision=mx.float32)
-    assert_allclose(paged_value, full_value, precision=mx.float32)
-
-
-def test_task_1_model_kv_caches_share_layer_pools():
-    mlx_model = _fake_qwen3_mlx_model()
-    week3_model = Qwen3ModelWeek3(mlx_model, page_size=4)
-    first_request_cache = week3_model.create_kv_cache()
-    second_request_cache = week3_model.create_kv_cache()
-
-    assert len(first_request_cache) == week3_model.num_hidden_layers
-    for layer in range(week3_model.num_hidden_layers):
-        assert first_request_cache[layer].pool is week3_model.page_pool
-        assert second_request_cache[layer].pool is week3_model.page_pool
-
-    assert first_request_cache[0].page_ids is not first_request_cache[1].page_ids
-    assert first_request_cache[0].page_lens is not first_request_cache[1].page_lens
-    assert first_request_cache[0].pool is first_request_cache[1].pool
-
-
-def test_task_1_model_layer_caches_keep_independent_page_metadata():
-    mlx_model = _fake_qwen3_mlx_model()
-    week3_model = Qwen3ModelWeek3(mlx_model, page_size=4)
-    cache = week3_model.create_kv_cache()
-    inputs = mx.array([[1, 5, 7, 3, 9]], dtype=mx.int32)
-
-    week3_model(inputs, 0, cache)
-
-    assert cache[0].page_ids == [0, 1]
-    assert cache[0].page_lens == [4, 1]
-    owned_page_ids = set(cache[0].page_ids)
-    for layer in range(1, week3_model.num_hidden_layers):
-        assert cache[layer].page_lens == cache[0].page_lens
-        assert owned_page_ids.isdisjoint(cache[layer].page_ids)
-        owned_page_ids.update(cache[layer].page_ids)
-        for page_id in cache[layer].page_ids:
-            key_page, value_page = week3_model.page_pool.read_page(page_id)
-            assert key_page.shape[2] == week3_model.page_size
-            assert value_page.shape[2] == week3_model.page_size
-
-
-def test_task_3_incremental_decode_matches_week2():
+def test_task_3_incremental_decode_matches_week2_with_paged_attention():
     mlx_model = _fake_qwen3_mlx_model()
     week2_model = Qwen3ModelWeek2(mlx_model)
     week3_model = Qwen3ModelWeek3(mlx_model, page_size=4)
@@ -234,5 +205,9 @@ def test_task_3_incremental_decode_matches_week2():
         week2_out = week2_out - mx.logsumexp(week2_out, keepdims=True)
         week3_out = week3_out - mx.logsumexp(week3_out, keepdims=True)
         assert_allclose(
-            week3_out, week2_out, precision=mx.bfloat16, rtol=0.1, atol=1.0
+            week3_out,
+            week2_out,
+            precision=mx.bfloat16,
+            rtol=0.1,
+            atol=1.0,
         )
