+++ tests_refsol/test_week_3_day_6.py
@@ -1,213 +1,133 @@
-"""Week 3 paged-attention runtime tests."""
+"""Optional Week 3 Day 6 MoE tests."""
 
 from types import SimpleNamespace
 
 import mlx.core as mx
+import mlx.nn as nn
+from mlx_lm.models.qwen3_moe import (
+    Qwen3MoeSparseMoeBlock as MlxQwen3MoeSparseMoeBlock,
+)
+from mlx_lm.models.switch_layers import SwitchLinear
 
 from .tiny_llm_base import (
-    BatchingKvCache,
-    Qwen3ModelWeek2,
-    Qwen3ModelWeek3,
-    TinyKvPagedCache,
-    TinyKvPagedPool,
-    flash_attention,
-    paged_attention,
-    scaled_dot_product_attention_grouped,
+    Moe,
+    QuantizedWeights,
+    grouped_expert_linear,
+    route_topk,
 )
 from .utils import assert_allclose
 
 
-def _random_chunk(
-    length: int, num_heads: int = 2, head_dim: int = 4
-) -> tuple[mx.array, mx.array]:
-    key = mx.random.normal(shape=(1, num_heads, length, head_dim)).astype(mx.float32)
-    value = mx.random.normal(shape=(1, num_heads, length, head_dim)).astype(mx.float32)
-    return key, value
-
-
-def _quantized_layer(
-    out_dim: int, in_dim: int, group_size: int = 128
-) -> SimpleNamespace:
-    weight = mx.random.normal(shape=(out_dim, in_dim), dtype=mx.bfloat16)
-    quantized_weight, scales, biases = mx.quantize(
-        weight, group_size=group_size, bits=4
-    )
-    return SimpleNamespace(
-        weight=quantized_weight,
-        scales=scales,
-        biases=biases,
-        group_size=group_size,
-        bits=4,
+def test_task_1_grouped_expert_linear():
+    mx.random.seed(1)
+    scale = 0.25
+    x = mx.random.normal(shape=(2, 3, 128), dtype=mx.float16) * scale
+    w_experts = mx.random.normal(shape=(3, 64, 128), dtype=mx.float16) * scale
+    expert_ids = mx.array(
+        [
+            [2, 0, 1],
+            [1, 2, 0],
+        ],
+        dtype=mx.uint32,
     )
 
-
-def _fake_qwen3_mlx_model() -> SimpleNamespace:
-    mx.random.seed(0)
-    args = SimpleNamespace(
-        num_hidden_layers=2,
-        hidden_size=128,
-        vocab_size=128,
-        num_attention_heads=4,
-        num_key_value_heads=2,
-        head_dim=32,
-        intermediate_size=256,
-        rms_norm_eps=1e-5,
-        max_position_embeddings=128,
-        rope_theta=10000,
-        tie_word_embeddings=True,
+    ref = SwitchLinear(
+        input_dims=w_experts.shape[-1],
+        output_dims=w_experts.shape[-2],
+        num_experts=w_experts.shape[0],
+        bias=False,
     )
-    embed_tokens = _quantized_layer(args.vocab_size, args.hidden_size)
-    kv_hidden_size = args.num_key_value_heads * args.head_dim
-    attn_hidden_size = args.num_attention_heads * args.head_dim
-    layers = []
-    for _ in range(args.num_hidden_layers):
-        layers.append(
-            SimpleNamespace(
-                self_attn=SimpleNamespace(
-                    q_proj=_quantized_layer(attn_hidden_size, args.hidden_size),
-                    k_proj=_quantized_layer(kv_hidden_size, args.hidden_size),
-                    v_proj=_quantized_layer(kv_hidden_size, args.hidden_size),
-                    o_proj=_quantized_layer(args.hidden_size, attn_hidden_size),
-                    q_norm=SimpleNamespace(
-                        weight=mx.ones((args.head_dim,), dtype=mx.bfloat16)
-                    ),
-                    k_norm=SimpleNamespace(
-                        weight=mx.ones((args.head_dim,), dtype=mx.bfloat16)
-                    ),
-                ),
-                mlp=SimpleNamespace(
-                    gate_proj=_quantized_layer(
-                        args.intermediate_size, args.hidden_size
-                    ),
-                    up_proj=_quantized_layer(args.intermediate_size, args.hidden_size),
-                    down_proj=_quantized_layer(
-                        args.hidden_size, args.intermediate_size
-                    ),
-                ),
-                input_layernorm=SimpleNamespace(
-                    weight=mx.ones((args.hidden_size,), dtype=mx.bfloat16)
-                ),
-                post_attention_layernorm=SimpleNamespace(
-                    weight=mx.ones((args.hidden_size,), dtype=mx.bfloat16)
-                ),
-            )
+    ref.weight = w_experts
+    ref = ref.to_quantized(group_size=128, bits=4)
+
+    out = grouped_expert_linear(
+        x,
+        QuantizedWeights.from_mlx_layer(ref),
+        expert_ids,
+    )
+    expected = ref(mx.expand_dims(x, -2), expert_ids).squeeze(-2)
+
+    assert out.shape == (2, 3, 64)
+    assert_allclose(out, expected, precision=mx.float16)
+
+
+def test_task_2_router_topk():
+    mx.random.seed(2)
+    scale = 0.25
+    x = mx.random.normal(shape=(2, 2, 128), dtype=mx.float16) * scale
+    ref = nn.Linear(128, 4, bias=False)
+    ref.weight = mx.random.normal(shape=(4, 128), dtype=mx.float16) * scale
+    ref = ref.to_quantized(group_size=128, bits=4)
+
+    router_probs, expert_ids, expert_scores = route_topk(
+        x,
+        QuantizedWeights.from_mlx_layer(ref),
+        top_k=2,
+    )
+    _, _, normalized_scores = route_topk(
+        x,
+        QuantizedWeights.from_mlx_layer(ref),
+        top_k=2,
+        norm_topk_prob=True,
+    )
+
+    expected_probs = mx.softmax(ref(x), axis=-1, precise=True)
+    expected_ids = mx.argpartition(-expected_probs, kth=1, axis=-1)[..., :2]
+    expected_scores = mx.take_along_axis(expected_probs, expected_ids, axis=-1)
+    expected_normalized_scores = expected_scores / expected_scores.sum(
+        axis=-1,
+        keepdims=True,
+    )
+
+    assert router_probs.shape == (2, 2, 4)
+    assert expert_ids.shape == (2, 2, 2)
+    assert expert_scores.shape == (2, 2, 2)
+    assert expert_ids.tolist() == expected_ids.tolist()
+    assert_allclose(router_probs, expected_probs, precision=mx.float16)
+    assert_allclose(expert_scores, expected_scores, precision=mx.float16)
+    assert_allclose(
+        normalized_scores,
+        expected_normalized_scores,
+        precision=mx.float16,
+    )
+
+
+def test_task_3_moe():
+    mx.random.seed(3)
+    scale = 0.25
+    x = mx.random.normal(shape=(2, 3, 128), dtype=mx.float16) * scale
+    ref = MlxQwen3MoeSparseMoeBlock(
+        SimpleNamespace(
+            hidden_size=128,
+            moe_intermediate_size=128,
+            num_experts=3,
+            num_experts_per_tok=2,
+            norm_topk_prob=True,
         )
-    return SimpleNamespace(
-        args=args,
-        model=SimpleNamespace(
-            embed_tokens=embed_tokens,
-            layers=layers,
-            norm=SimpleNamespace(
-                weight=mx.ones((args.hidden_size,), dtype=mx.bfloat16)
-            ),
-        ),
+    )
+    ref.gate.weight = mx.random.normal(shape=(3, 128), dtype=mx.float16) * scale
+    ref.switch_mlp.gate_proj.weight = (
+        mx.random.normal(shape=(3, 128, 128), dtype=mx.float16) * scale
+    )
+    ref.switch_mlp.up_proj.weight = (
+        mx.random.normal(shape=(3, 128, 128), dtype=mx.float16) * scale
+    )
+    ref.switch_mlp.down_proj.weight = (
+        mx.random.normal(shape=(3, 128, 128), dtype=mx.float16) * scale
+    )
+    nn.quantize(ref, group_size=128, bits=4)
+
+    moe = Moe(
+        w_router=QuantizedWeights.from_mlx_layer(ref.gate),
+        w_gate=QuantizedWeights.from_mlx_layer(ref.switch_mlp.gate_proj),
+        w_up=QuantizedWeights.from_mlx_layer(ref.switch_mlp.up_proj),
+        w_down=QuantizedWeights.from_mlx_layer(ref.switch_mlp.down_proj),
+        num_experts_per_tok=2,
+        norm_topk_prob=True,
     )
 
+    out = moe(x)
+    expected = ref(x)
 
-def test_task_1_paged_attention_matches_dense_flash_attention():
-    page_size = 4
-    pool = TinyKvPagedPool(page_size=page_size)
-    cache = TinyKvPagedCache(pool=pool)
-    first_key, first_value = _random_chunk(3)
-    second_key, second_value = _random_chunk(3)
-
-    cache.update_and_fetch(first_key, first_value)
-    metadata = cache.update_and_fetch_paged(second_key, second_value, mask="causal")
-
-    query = mx.random.normal(shape=(1, 4, second_key.shape[2], 4)).astype(mx.float32)
-    dense_key, dense_value = cache.gather_dense()
-    dense_output = flash_attention(
-        query,
-        dense_key,
-        dense_value,
-        mask="causal",
-    )
-    paged_output = paged_attention(
-        query,
-        metadata.key_pages,
-        metadata.value_pages,
-        metadata.block_table,
-        metadata.context_lens,
-        metadata.page_size,
-        mask=metadata.mask,
-    )
-
-    assert metadata.block_table.tolist() == [[0, 1]]
-    assert metadata.context_lens.tolist() == [6]
-    assert metadata.key_pages.shape == (2, 2, page_size, 4)
-    assert metadata.value_pages.shape == (2, 2, page_size, 4)
-    assert_allclose(paged_output, dense_output, precision=mx.float32)
-
-
-def test_task_2_batched_paged_attention_matches_dense_attention():
-    page_size = 4
-    pool = TinyKvPagedPool(page_size=page_size)
-    first = TinyKvPagedCache(pool=pool)
-    second = TinyKvPagedCache(pool=pool)
-    first.update_and_fetch(*_random_chunk(3))
-    second.update_and_fetch(*_random_chunk(6))
-
-    batch = BatchingKvCache(max_active_requests=3, max_seq_len=16)
-    batch.add_request(first, 0)
-    batch.add_request(second, 2)
-
-    keys = mx.zeros((3, 2, 1, 4), dtype=mx.float32)
-    values = mx.zeros((3, 2, 1, 4), dtype=mx.float32)
-    keys[0:1], values[0:1] = _random_chunk(1)
-    keys[2:3], values[2:3] = _random_chunk(1)
-
-    metadata = batch.update_and_fetch_paged(
-        keys,
-        values,
-        mask_length=1,
-        mask="causal",
-    )
-    query = mx.random.normal(shape=(3, 4, 1, 4)).astype(mx.float32)
-    paged_output = paged_attention(
-        query,
-        metadata.key_pages,
-        metadata.value_pages,
-        metadata.block_table,
-        metadata.context_lens,
-        metadata.page_size,
-        mask=metadata.mask,
-    )
-
-    first_key, first_value = first.gather_dense()
-    first_output = scaled_dot_product_attention_grouped(
-        query[0:1], first_key, first_value, mask="causal"
-    )
-    second_key, second_value = second.gather_dense()
-    second_output = scaled_dot_product_attention_grouped(
-        query[2:3], second_key, second_value, mask="causal"
-    )
-
-    assert metadata.context_lens.tolist() == [4, 0, 7]
-    assert metadata.block_table.shape == (3, 2)
-    assert metadata.block_table.tolist()[1] == [-1, -1]
-    assert metadata.key_pages.shape == (3, 2, page_size, 4)
-    assert_allclose(paged_output[0:1], first_output, precision=mx.float32)
-    assert_allclose(paged_output[2:3], second_output, precision=mx.float32)
-
-
-def test_task_3_incremental_decode_matches_week2_with_paged_attention():
-    mlx_model = _fake_qwen3_mlx_model()
-    week2_model = Qwen3ModelWeek2(mlx_model)
-    week3_model = Qwen3ModelWeek3(mlx_model, page_size=4)
-    inputs = mx.array([[1, 5, 7, 3, 9, 11]], dtype=mx.int32)
-    week2_cache = week2_model.create_kv_cache()
-    week3_cache = week3_model.create_kv_cache()
-
-    for offset in range(inputs.shape[1]):
-        token = inputs[:, offset : offset + 1]
-        week2_out = week2_model(token, offset, week2_cache)
-        week3_out = week3_model(token, offset, week3_cache)
-        week2_out = week2_out - mx.logsumexp(week2_out, keepdims=True)
-        week3_out = week3_out - mx.logsumexp(week3_out, keepdims=True)
-        assert_allclose(
-            week3_out,
-            week2_out,
-            precision=mx.bfloat16,
-            rtol=0.1,
-            atol=1.0,
-        )
+    assert out.shape == x.shape
+    assert_allclose(out, expected, precision=mx.float16)
